warning C4244: conversion from '__int64' to 'unsigned int'
The difference between `file' and `flash_RAM' is always within the guaranteed range of values that will fit within any implementation size of type `int', so this warning can be disposed of safely by means of an explicit cast.

+++ main.c
@@ -28,7 +28,7 @@ int main(int argc, char ** argv)
     while (i < argc)
         i += opt_execute(&argv[i]);
 
-    i = (file - flash_RAM) / FILE_SIZE;
+    i = (int)(file - flash_RAM) / FILE_SIZE;
     printf(
         "Saved checksum 0x%04X to section %i.\n",
         fix_checksum(i), i
